Fixed #75 + Add generic configurations for product price + inventory import + Add generic LastEntityAndRowIdObserver that loads the product by the SKU found in the CSV file and set the row ID as lastRowId
