+++ etc/techdivision-import-inventory.json
@@ -1,0 +1,83 @@
+{
+  "magento-edition": "EE",
+  "magento-version": "2.1.2",
+  "operation-name" : "add-update",
+  "source-date-format": "n/d/y, g:i A",
+  "archive-artefacts" : true,
+  "debug-mode" : false,
+  "entity-type-code" : "catalog_product_inventory",
+  "system-name" : "system-name/product-import-inventory",
+  "databases" : [
+  ],
+  "loggers": [
+    {
+      "name": "system",
+      "channel-name": "logger/system",
+      "type": "Monolog\\Logger",
+      "handlers": [
+        {
+          "type": "Monolog\\Handler\\ErrorLogHandler",
+          "formatter": {
+            "type": "Monolog\\Formatter\\LineFormatter",
+            "params" : [
+              {
+                "format": "[%datetime%] %channel%.%level_name%: %message% %context% %extra%",
+                "dateFormat": "Y-m-d H:i:s",
+                "allowInlineLineBreaks": true,
+                "ignoreEmptyContextAndExtra": false
+              }
+            ]
+          }
+        }
+      ]
+    }
+  ],
+  "operations" : [
+    {
+      "name" : "add-update",
+      "plugins" : [
+        {
+          "id": "import.plugin.global.data"
+        },
+        {
+          "id": "import.plugin.subject",
+          "subjects": [
+            {
+              "id": "import.subject.move.files",
+              "identifier": "move-files",
+              "prefix": "product-import-inventory",
+              "ok-file-needed": true
+            },
+            {
+              "id": "import_product.subject.bunch",
+              "identifier": "files",
+              "prefix": "product-import-inventory",
+              "observers": [
+                {
+                  "pre-import": [
+                    "import.observer.attribute.set",
+                    "import_product.observer.last.entity.id",
+                    "import_product.observer.quality.and.stock.status"
+                  ]
+                },
+                {
+                  "import": [
+                    "import_product.observer.product.inventory.update"
+                  ]
+                },
+                {
+                  "post-import": [
+                    "import_product.observer.clean.up"
+                  ]
+                }
+              ]
+            }
+          ]
+        },
+        {
+          "id": "import.plugin.archive"
+        }
+      ]
+    }
+  ]
+}
