+++ etc/techdivision-import-price.json
@@ -1,0 +1,82 @@
+{
+  "magento-edition": "CE",
+  "magento-version": "2.1.2",
+  "operation-name" : "add-update",
+  "source-date-format": "n/d/y, g:i A",
+  "archive-artefacts" : true,
+  "debug-mode" : false,
+  "entity-type-code" : "catalog_product_price",
+  "system-name" : "system-name/product-import-price",
+  "databases" : [
+  ],
+  "loggers": [
+    {
+      "name": "system",
+      "channel-name": "logger/system",
+      "type": "Monolog\\Logger",
+      "handlers": [
+        {
+          "type": "Monolog\\Handler\\ErrorLogHandler",
+          "formatter": {
+            "type": "Monolog\\Formatter\\LineFormatter",
+            "params" : [
+              {
+                "format": "[%datetime%] %channel%.%level_name%: %message% %context% %extra%",
+                "dateFormat": "Y-m-d H:i:s",
+                "allowInlineLineBreaks": true,
+                "ignoreEmptyContextAndExtra": false
+              }
+            ]
+          }
+        }
+      ]
+    }
+  ],
+  "operations" : [
+    {
+      "name" : "add-update",
+      "plugins" : [
+        {
+          "id": "import.plugin.global.data"
+        },
+        {
+          "id": "import.plugin.subject",
+          "subjects": [
+            {
+              "id": "import.subject.move.files",
+              "identifier": "move-files",
+              "prefix": "product-import-price",
+              "ok-file-needed": true
+            },
+            {
+              "id": "import_product_ee.subject.bunch",
+              "identifier": "files",
+              "prefix": "product-import-price",
+              "observers": [
+                {
+                  "pre-import": [
+                    "import.observer.attribute.set",
+                    "import_product_ee.observer.last.entity.and.row.id"
+                  ]
+                },
+                {
+                  "import": [
+                    "import_product_ee.observer.product.attribute.update"
+                  ]
+                },
+                {
+                  "post-import": [
+                    "import_product.observer.clean.up"
+                  ]
+                }
+              ]
+            }
+          ]
+        },
+        {
+          "id": "import.plugin.archive"
+        }
+      ]
+    }
+  ]
+}
