+++ src/Observers/EeProductObserver.php
@@ -22,6 +22,7 @@ namespace TechDivision\Import\Product\Ee\Observers;
 
 use TechDivision\Import\Product\Ee\Utils\MemberNames;
 use TechDivision\Import\Product\Observers\ProductObserver;
+use TechDivision\Import\Utils\EntityStatus;
 
 /**
  * Observer that create's the product itself for the Magento 2 EE edition.
@@ -61,15 +62,24 @@ class EeProductObserver extends ProductObserver
     protected function initializeProduct(array $attr)
     {
 
-        // initialize the addtional Magento 2 EE product values
-        $additionalAttr = array(
-            MemberNames::ENTITY_ID  => $this->nextIdentifier(),
-            MemberNames::CREATED_IN => 1,
-            MemberNames::UPDATED_IN => strtotime('+20 years')
-        );
+        // initialize the product attributes
+        $attr = parent::initializeProduct($attr);
 
-        // merge and return the attributes
-        return array_merge($attr, $additionalAttr);
+        // query whether or not, we found a new product
+        if ($attr[EntityStatus::MEMBER_NAME] === EntityStatus::STATUS_CREATE) {
+            // if yes, initialize the additional Magento 2 EE product values
+            $additionalAttr = array(
+                MemberNames::ENTITY_ID  => $this->nextIdentifier(),
+                MemberNames::CREATED_IN => 1,
+                MemberNames::UPDATED_IN => strtotime('+20 years')
+            );
+
+            // merge and return the attributes
+            $attr = array_merge($attr, $additionalAttr);
+        }
+
+        // otherwise simply return the attributes
+        return $attr;
     }
 
     /**
