+++ src/Observers/LastEntityAndRowIdObserver.php
@@ -1,0 +1,66 @@
+<?php
+
+/**
+ * TechDivision\Import\Product\Observers\LastEntityAndRowIdObserver
+ *
+ * NOTICE OF LICENSE
+ *
+ * This source file is subject to the Open Software License (OSL 3.0)
+ * that is available through the world-wide-web at this URL:
+ * http://opensource.org/licenses/osl-3.0.php
+ *
+ * PHP version 5
+ *
+ * @author    Tim Wagner <t.wagner@techdivision.com>
+ * @copyright 2016 TechDivision GmbH <info@techdivision.com>
+ * @license   http://opensource.org/licenses/osl-3.0.php Open Software License (OSL 3.0)
+ * @link      https://github.com/techdivision/import-product
+ * @link      http://www.techdivision.com
+ */
+
+namespace TechDivision\Import\Product\Ee\Observers;
+
+use TechDivision\Import\Product\Ee\Utils\MemberNames;
+use TechDivision\Import\Product\Observers\LastEntityIdObserver;
+
+/**
+ * Observer that pre-loads the entity ID of the product with the SKU found in the CSV file.
+ *
+ * @author    Tim Wagner <t.wagner@techdivision.com>
+ * @copyright 2016 TechDivision GmbH <info@techdivision.com>
+ * @license   http://opensource.org/licenses/osl-3.0.php Open Software License (OSL 3.0)
+ * @link      https://github.com/techdivision/import-product
+ * @link      http://www.techdivision.com
+ */
+class LastEntityAndRowIdObserver extends LastEntityIdObserver
+{
+
+    /**
+     * Temporarily persist's the IDs of the passed product.
+     *
+     * @param array $product The product to temporarily persist the IDs for
+     *
+     * @return void
+     */
+    protected function setIds(array $product)
+    {
+
+        // pass the product to the parent method
+        parent::setIds($product);
+
+        // temporarily persist the row ID
+        $this->setLastRowId($product[MemberNames::ROW_ID]);
+    }
+
+    /**
+     * Set's the row ID of the product that has been created recently.
+     *
+     * @param string $rowId The row ID
+     *
+     * @return void
+     */
+    protected function setLastRowId($rowId)
+    {
+        $this->getSubject()->setLastRowId($rowId);
+    }
+}
